Update ScenarioAnalyze.py
Adding MergeSavedFiles option to Disabling Recorders._MergeAndClear() for objs_recorder, because it took a long time for huge models and great number of analysis. By this option users can determine the resulted files merge to one file or not.

Also, using MaxExistSavedfileIndex variable the number of saving files in parallel analysis are checked and if other saved files exist, the filenam number will be started after the maximum saved number.