Update Objs_recorder.py
IMPORTANT ENHANCEMENT of OBJS_RECORDER to manage memory:

RecodingSubpackages option added for saving just object of subpackages that user need to manage memory and prevent memory loss.

OtherSaveOnce method is for saving subpackages that are not defined by user for saving in each analysis or simulation and using this method are saved just once.

Also, if user wanna merge all saved files in 1 file, the merge method will save all files in a file with name "filename+M.OPR" and "filename+M.Log"

LoadOtherSubPackages method is for loading other subpackages and this is also in the control of user and user should decide to load them or not.